refactor(usecase): rename file to keep pattern on naming
