+++ src/core/usecase/GenerateKinesisEvents.ts
